% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanmachine-package.R
\docType{package}
\name{cleanmachine-package}
\alias{cleanmachine}
\alias{cleanmachine-package}
\title{cleanmachine: Clean Variables Using Rules from a Data File}
\description{
Clean lots of survey variables using a CSV file to define cleaning rules.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/davidskalinder/cleanmachine}
  \item Report bugs at \url{https://github.com/davidskalinder/cleanmachine/issues}
}

}
\keyword{internal}
