% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gss_demo}
\alias{gss_demo}
\title{A sample from the 2004 GSS.}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 100 rows and 13 columns.
}
\usage{
gss_demo
}
\description{
A sample from the 2004 GSS.
}
\keyword{datasets}
