% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{write_metadata_skeleton}
\alias{write_metadata_skeleton}
\title{Write a skeleton metadata file}
\usage{
write_metadata_skeleton(
  source_vars,
  outfile,
  source_varname_col,
  num_exprs = 5
)
}
\arguments{
\item{source_vars}{A data frame with one row per source variable.}

\item{outfile}{Destination of the skeleton file.}

\item{source_varname_col}{Tidy-select column in \code{source_vars} with the
name of the variable.}

\item{num_exprs}{Number of new/old expression cols to include in file.}
}
\value{
Nothing, called only for side effects.
}
\description{
Write a skeleton metadata file
}
\examples{
write_metadata_skeleton(
  source_vars = tibble::tibble(source_varname = names(gss_demo)),
  outfile = tempfile(),
  source_varname_col = source_varname,
  num_exprs = 5
)
}
