% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning_functions.R
\name{apply_cleaning_rules}
\alias{apply_cleaning_rules}
\title{Apply cleaning rules}
\usage{
apply_cleaning_rules(dataset, varinfo, type_fn_prefix = "as.")
}
\arguments{
\item{dataset}{A dataset}

\item{varinfo}{Variable info for the dataset}

\item{type_fn_prefix}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}} Prefix for cleaning_data_type and final_data_type cols}
}
\value{
The cleaned dataset.
}
\description{
Note that \code{cleaning_data_type} is applied to each column before \emph{and}
after cleaning.  Then \code{final_data_type} is applied just before output.
This behavior should be restructured in future versions.
}
